web/requirements.txt
Flask-HTTPAuth==3.2.3
Flask==1.0.2
gspread==3.0.0
oauth2client==4.1.2
python-dateutil==2.7.5
pytz==2018.7
